--- a/quic/s2n-quic-bench/src/lib.rs
+++ b/quic/s2n-quic-bench/src/lib.rs
@@ -3,12 +3,14 @@
 
 use criterion::Criterion;
 
+mod buffer;
 mod crypto;
 mod frame;
 mod packet;
 mod varint;
 
 pub fn benchmarks(c: &mut Criterion) {
+    buffer::benchmarks(c);
     crypto::benchmarks(c);
     frame::benchmarks(c);
     packet::benchmarks(c);