--- a/quic/s2n-quic-core/src/buffer/mod.rs
+++ b/quic/s2n-quic-core/src/buffer/mod.rs
@@ -4,6 +4,3 @@
 mod receive_buffer;
 
 pub use receive_buffer::*;
-
-#[cfg(test)]
-mod tests;