--- a/quic/s2n-quic-core/src/buffer/receive_buffer/__fuzz__/buffer__receive_buffer__tests__model/corpus.tar.gz
+++ b/quic/s2n-quic-core/src/buffer/receive_buffer/__fuzz__/buffer__receive_buffer__tests__model/corpus.tar.gz
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:50a8a3411017ead24ca691242c57e0231f37c17a5eabcfb18c5d2d354607b70a
+size 3225600