--- a/quic/s2n-quic-core/src/lib.rs
+++ b/quic/s2n-quic-core/src/lib.rs
@@ -8,6 +8,8 @@ extern crate alloc;
 
 pub mod ack;
 pub mod application;
+#[cfg(feature = "alloc")]
+pub mod buffer;
 pub mod connection;
 pub mod counter;
 pub mod crypto;