### Description of changes: 

On main, the receive buffer occupies about ~20% of CPU cycles when receiving a lot of data. This PR optimizes the receive buffer and reduces it to ~4%.

#### Benchmarks

| frame size (bytes) | throughput (before) | throughput (after) | speedup |
|--------------------|---------------------|--------------------|---------|
| 1                  | 10 MiB/s            | 37 MiB/s           | 3.7x    |
| 100                | 355 MiB/s           | 1.3 GiB/s          | 3.6x    |
| 1000               | 1015 MiB/s          | 2.264 GiB/s        | 2.2x    |
| 1450               | 1.338 GiB/s         | 2.299 GiB/s        | 1.7x    |

You can run the benchmarks in the `s2n-quic-bench` crate:

```
cd quic/s2n-quic-bench
cargo bench buffer
```

A few optimizations that went into this refactor:

#### Larger allocations

The previous version always allocated 4k pages in order to reduce the amount of memory held on to in each of the received chunks. However, this results in a lot of interaction with the global allocator. The new version, instead, gradually increases the allocation sizes as the offset increases. This is under the assumption that streams that receive a lot of data will continue to receive a lot of data.

| bytes received | allocation size |
|----------------|-----------------|
| 0              | 4096            |
| 65536          | 16384           |
| 262144         | 32768           |
| >=1048575      | 65536           |

#### Reverse iteration

The previous version iterated over chunks from low offset to high. This isn't ideal since most chunks will be coming in order and will just extend the highest received offset. The new implementation, instead, iterates from high to low, which greatly helps the ideal scenario.

#### Optimized unsplit

The previous version had a lot of checks in place to unsplit chunks after receiving, which checked both left and right allocations. The new version only checks left allocations because it iterates over the queue in reverse.

#### Before
[
![](https://dnglbrstg7yg.cloudfront.net/1645cc5ee2f3fde753cf47d881bcdccd6648639d/perf/s2n-quic-client-s2n-quic-server/1000MB-down-0MB-up.client.svg)
](https://dnglbrstg7yg.cloudfront.net/1645cc5ee2f3fde753cf47d881bcdccd6648639d/perf/s2n-quic-client-s2n-quic-server/1000MB-down-0MB-up.client.svg)

#### After

[
![](https://dnglbrstg7yg.cloudfront.net/a71a192a45c55dccd8e9b3d7749087803ca8eb23/perf/s2n-quic-client-s2n-quic-server/1000MB-down-0MB-up.client.svg)
](https://dnglbrstg7yg.cloudfront.net/a71a192a45c55dccd8e9b3d7749087803ca8eb23/perf/s2n-quic-client-s2n-quic-server/1000MB-down-0MB-up.client.svg)

### Call-outs:

* I've moved the receive buffer into `s2n-quic-core` to make it easier to add benchmarks for

### Testing:

All of the existing tests were preserved from the previous version. I've also added a fuzz test and some internal consistency checks to improve the testing coverage.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.