@camshaft 2022-09-21
I added some instructions to run the benchmarks. For the flamgraphs I'm just linking to the perf artifacts for this PR: https://dnglbrstg7yg.cloudfront.net/026edf713ff2f60c35e15021e25c9006084940af/perf/index.html.